Refactor telemetry logic and show values
