+++ frontend/static/style.css
@@ -1,3 +1,9 @@
 body {
   font-family: 'Roboto', 'Segoe UI', Tahoma, sans-serif;
 }
+
+.telemetry-value {
+  font-size: 2rem;
+  font-weight: bold;
+  text-align: center;
+}
