+++ frontend/templates/v2/pi-telemetry.html
@@ -4,15 +4,19 @@
 <h1 class="mb-4">Raspberry Pi Telemetry</h1>
 <div class="row g-4">
   <div class="col-md-6">
+    <h2 class="telemetry-value" id="cpuTempValue">--</h2>
     <canvas id="cpuTempChart" height="200"></canvas>
   </div>
   <div class="col-md-6">
+    <h2 class="telemetry-value" id="cpuFreqValue">--</h2>
     <canvas id="cpuFreqChart" height="200"></canvas>
   </div>
   <div class="col-md-6">
+    <h2 class="telemetry-value" id="memValue">--</h2>
     <canvas id="memChart" height="200"></canvas>
   </div>
   <div class="col-md-6">
+    <h2 class="telemetry-value" id="diskValue">--</h2>
     <canvas id="diskChart" height="200"></canvas>
   </div>
 </div>
